Move BeamConfig and DetectConfig to separate files
